requirements.txt
bidict==0.22.1
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-PyMongo==2.3.0
greenlet==3.0.1
h11==0.14.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.5.0
python-dotenv==1.0.0
python-engineio==4.8.0
simple-websocket==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.17.0
